WIP: core interceptor refactor.
This starts a refactor to support "core" event interceptors. This moves
the existing "ObjectRef" style interceptor to a new package called "webhook",
and embeds it as a "WebhookInterceptor" on the EventBinding type.

Other interceptors (Github, Gitlab, etc.) will be added alongside this. This
PR also extracts the beginnings of the "Interceptor" interface that these other
interceptors will implement.